fix(gitgraph): 跨 lane 过渡改用 dot-to-dot 紧凑 S 形 (3 段 L)
vscode 原版用 C 贝塞尔 + d = grid.y * 0.8 = 19.2px, 控制点偏移
超出整行高, 曲线会拉得"很缓", 弯折几乎占满整行 (24px)。

按用户描述"充分利用行间距, 快速完成曲线转向, 在 row1 底→
row2 顶 的小空间内完成", 改为 3 段 L 命令:
  - dot 中心 (x1, y1)
  - 沿 x 走到 row1 底 midY1 = y1 + GRID_Y/2 - dy
  - 横移到 x2 (midY1 高度)
  - 垂直降到 row2 顶 midY2 = y2 - GRID_Y/2 + dy
  - 沿 x 走到 dot 中心 (x2, y2)

dy ≈ dot 半径 (3px), 整个弯折在 [midY1, midY2] = 6px 空间内完成,
看起来像 dot 之间的紧凑"小弯", 而不是 24px 的"大弧"。

同时修正 OFFSET_X=16, OFFSET_Y=12 (vscode src/config.ts:278
默认值, 不是之前复刻时猜的 4)。

涉及文件:
  - tools/vscode-recheck/debug/main.go: 同策略, 生成 HTML 验证
  - frontend/src/lib/gitgraph/vscode-render.ts: 实际渲染用
  - frontend/src/lib/gitgraph/__tests__/vscode_render.test.ts: 更新断言
    (10 个 vitest case 全 PASS)

+++ tools/vscode-recheck/debug/main.go
@@ -234,9 +234,9 @@ func renderGraphVscode(g *graph.GraphResult) ([]pathOut, []nodeOut) {
 				// dot-to-dot 紧凑 S 形过渡: 在 row1 底→row2 顶 的小空间内完成
 				// (vscode 原版 C 贝塞尔 d=0.8*GRID_Y=19.2 拉得太缓, 跟"dot 之间紧凑过渡"不符)
 				// 策略: dot 中心 (x1,y1) → row1 底部 midY1 沿 x 走到 x2 → 垂直降到 row2 顶部 midY2 → 沿 x 到 (x2,y2)
-				dy := CURVE_CONTROL_DY
-				midY1 := y1 + GRID_Y/2 - dy
-				midY2 := y2 - GRID_Y/2 + dy
+				dy := float64(CURVE_CONTROL_DY)
+				midY1 := y1 + float64(GRID_Y)/2 - dy
+				midY2 := y2 - float64(GRID_Y)/2 + dy
 				cur += fmt.Sprintf(" L %.0f %.1f L %.0f %.1f L %.0f %.1f",
 					x1, midY1, x2, midY1, x2, midY2)
 			}
